// client/src/constants/todoConstants.ts
export enum TodoConstants {
  GET_TODOS_LOADING = 'GET_TODOS_LOADING',
  GET_TODOS_SUCCESS = 'GET_TODOS_SUCCESS',
  GET_TODOS_FAIL = 'GET_TODOS_FAIL',
  ADD_TODO_LOADING = 'ADD_TODO_LOADING',
  ADD_TODO_SUCCESS = 'ADD_TODO_SUCCESS',
  ADD_TODO_FAIL = 'ADD_TODO_FAIL',
  UPDATE_TODO_LOADING = 'UPDATE_TODO_LOADING',
  UPDATE_TODO_SUCCESS = 'UPDATE_TODO_SUCCESS',
  UPDATE_TODO_FAIL = 'UPDATE_TODO_FAIL',
  DELETE_TODO_LOADING = 'DELETE_TODO_LOADING',
  DELETE_TODO_SUCCESS = 'DELETE_TODO_SUCCESS',
  DELETE_TODO_FAIL = 'DELETE_TODO_FAIL',
}